update fops + docs
